move dir UserController.php
move UserController from /controllers/ to /controllers/api/v1

corrects namespaces and paths
