+++ app/Http/Controllers/Api/V1/UserController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api\V1;
 
+use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
@@ -20,6 +21,7 @@ class UserController extends Controller
                 'user'=> $user,
                 'token'=> $token
             ];
+
             return response($response, 201);
     }
 }
